2010-04-25	Joel Sherrill <joel.sherrilL@OARcorp.com>
	* libchip/serial/mc68681.c, libchip/serial/mc68681_reg.c,
	libchip/serial/ns16550.c, libchip/serial/serial.h,
	libchip/serial/z85c30.c, libchip/serial/z85c30_reg.c: Use uintptr_t
	or intptr_t instead of uint32_t or int32_t.

+++ c/src/libchip/serial/mc68681_reg.c
@@ -31,7 +31,7 @@
  */
 
 uint8_t   _MC68681_NAME(mc68681_get_register)(
-  uint32_t    ulCtrlPort,
+  uintptr_t   ulCtrlPort,
   uint8_t     ucRegNum
 )
 {
@@ -47,7 +47,7 @@ uint8_t   _MC68681_NAME(mc68681_get_register)(
  */
 
 void  _MC68681_NAME(mc68681_set_register)(
-  uint32_t    ulCtrlPort,
+  uintptr_t   ulCtrlPort,
   uint8_t     ucRegNum,
   uint8_t     ucData
 )
